--- src/main/java/com/covelopfit/autotrading/controller/OrderForm.java
@@ -1,49 +1,0 @@
-package com.covelopfit.autotrading.controller;
-
-public class OrderForm {
-    private String market;
-    private String side;
-    private String volume;
-    private String price;
-    private String ord_type;
-
-    public String getMarket() {
-        return market;
-    }
-
-    public void setMarket(String market) {
-        this.market = market;
-    }
-
-    public String getSide() {
-        return side;
-    }
-
-    public void setSide(String side) {
-        this.side = side;
-    }
-
-    public String getVolume() {
-        return volume;
-    }
-
-    public void setVolume(String volume) {
-        this.volume = volume;
-    }
-
-    public String getPrice() {
-        return price;
-    }
-
-    public void setPrice(String price) {
-        this.price = price;
-    }
-
-    public String getOrd_type() {
-        return ord_type;
-    }
-
-    public void setOrd_type(String ord_type) {
-        this.ord_type = ord_type;
-    }
-}
