--- src/main/java/com/covelopfit/autotrading/domain/UpbitKey.java
@@ -1,22 +1,0 @@
-package com.covelopfit.autotrading.domain;
-
-public class UpbitKey {
-    private String accessKey;
-    private String secretKey;
-
-    public String getAccessKey() {
-        return accessKey;
-    }
-
-    public void setAccessKey(String accessKey) {
-        this.accessKey = accessKey;
-    }
-
-    public String getSecretKey() {
-        return secretKey;
-    }
-
-    public void setSecretKey(String secretKey) {
-        this.secretKey = secretKey;
-    }
-}
