+++ src/main/resources/static/index.html
@@ -5,6 +5,6 @@
   <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" /> </head>
 <body>
 Hello
-<a href="/order-page">order</a>
+<a href="/hello">hello</a>
 </body>
 </html>
