+++ src/main/resources/templates/order-page.html
@@ -6,7 +6,7 @@
 </head>
 <body>
     order-page
-    <form action="/buy-order" method="post"> <div class="form-group">
+    <form action="/order" method="post"> <div class="form-group">
         <label for="market">코인 종류</label>
         <input type="text" id="market" name="market" placeholder="이름을 입력하세요">
         <label for="side">주문 종류</label>
